--- src/grain/createGrainDistribution.test.js
@@ -1,490 +1,0 @@
-// @flow
-
-import {NodeAddress} from "../core/graph";
-import {ONE, ZERO, fromApproximateFloat, format} from "./grain";
-import {
-  GRAIN_DISTRIBUTION_VERSION_1,
-  createGrainDistribution,
-} from "./createGrainDistribution";
-import deepFreeze from "deep-freeze";
-import type {
-  CredHistory,
-  BalancedV1,
-  ImmediateV1,
-  GrainReceipt,
-  DistributionStrategy,
-  GrainDistributionV1,
-} from "./createGrainDistribution";
-import type {Grain} from "./grain";
-
-describe("src/grain/createGrainDistribution", () => {
-  const foo = NodeAddress.fromParts(["foo"]);
-  const bar = NodeAddress.fromParts(["bar"]);
-
-  const timestampMs = 1000;
-
-  const credHistory: CredHistory = deepFreeze([
-    {
-      intervalEndMs: 10,
-      cred: new Map([
-        [foo, 9],
-        [bar, 1],
-      ]),
-    },
-    {
-      intervalEndMs: 20,
-      cred: new Map([
-        [foo, 1],
-        [bar, 1],
-      ]),
-    },
-    {intervalEndMs: 30, cred: new Map([[bar, 2]])},
-  ]);
-
-  // there's no JSON serialization for BigInts, so we need to convert
-  // the BigInts before passing them into expect comparisons. Otherwise,
-  // when the tests fail, they'll fail with an unhelpful JSON serialization
-  // issue.
-  function fmt(g: Grain) {
-    return format(g, 3);
-  }
-  function safeReceipts(receipts: $ReadOnlyArray<GrainReceipt>) {
-    return receipts.map(({address, amount}) => ({
-      address,
-      amount: fmt(amount),
-    }));
-  }
-  function safeStrategy(strat: DistributionStrategy) {
-    return {...strat, budget: fmt(strat.budget)};
-  }
-  function safeDistribution(distribution: GrainDistributionV1) {
-    return {
-      ...distribution,
-      strategy: safeStrategy(distribution.strategy),
-      receipts: safeReceipts(distribution.receipts),
-    };
-  }
-  function expectDistributionsEqual(h1, h2) {
-    expect(safeDistribution(h1)).toEqual(safeDistribution(h2));
-  }
-
-  describe("immediateDistribution", () => {
-    const strategy: ImmediateV1 = deepFreeze({
-      type: "IMMEDIATE",
-      budget: ONE,
-      version: 1,
-    });
-
-    describe("it should return an empty createGrainDistribution when", () => {
-      const emptyDistribution = deepFreeze({
-        type: "GRAIN_DISTRIBUTION",
-        version: GRAIN_DISTRIBUTION_VERSION_1,
-        receipts: [],
-        strategy,
-        timestampMs,
-      });
-
-      it("the budget is zero", () => {
-        const zeroBudgetStrategy = {...strategy, budget: ZERO};
-
-        const actual = createGrainDistribution(
-          zeroBudgetStrategy,
-          credHistory,
-          new Map(),
-          timestampMs
-        );
-
-        expectDistributionsEqual(actual, {
-          ...emptyDistribution,
-          strategy: zeroBudgetStrategy,
-        });
-      });
-
-      it("there are no cred scores at all", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          [],
-          new Map(),
-          timestampMs
-        );
-        expectDistributionsEqual(actual, emptyDistribution);
-      });
-
-      it("all cred scores are from the future", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          credHistory,
-          new Map(),
-          0
-        );
-        expectDistributionsEqual(actual, {
-          ...emptyDistribution,
-          timestampMs: 0,
-        });
-      });
-
-      it("all the cred sums to 0", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          [
-            {
-              intervalEndMs: timestampMs - 500,
-              cred: new Map([
-                [foo, 0],
-                [bar, 0],
-              ]),
-            },
-          ],
-          new Map(),
-          timestampMs
-        );
-
-        expectDistributionsEqual(actual, emptyDistribution);
-      });
-    });
-
-    const createImmediateDistribution = (
-      timestampMs: number,
-      receipts: $ReadOnlyArray<GrainReceipt>
-    ) => {
-      return {
-        type: "GRAIN_DISTRIBUTION",
-        version: GRAIN_DISTRIBUTION_VERSION_1,
-        receipts,
-        strategy,
-        timestampMs,
-      };
-    };
-
-    it("throws an error if given an unsupported strategy", () => {
-      const unsupportedStrategy = {
-        ...strategy,
-        version: 2,
-      };
-      expect(() =>
-        createGrainDistribution(
-          unsupportedStrategy,
-          credHistory,
-          new Map(),
-          timestampMs
-        )
-      ).toThrowError(`Unsupported IMMEDIATE strategy: 2`);
-    });
-
-    it("throws an error if given an unsupported strategy and an empty credHistory", () => {
-      const unsupportedStrategy = {
-        ...strategy,
-        version: 2,
-      };
-      expect(() =>
-        createGrainDistribution(unsupportedStrategy, [], new Map(), timestampMs)
-      ).toThrowError(`Unsupported IMMEDIATE strategy: 2`);
-    });
-
-    it("handles an interval in the middle", () => {
-      const result = createGrainDistribution(
-        strategy,
-        credHistory,
-        new Map(),
-        20
-      );
-      // $ExpectFlowError
-      const HALF = ONE / 2n;
-      const expectedReceipts = [
-        {address: foo, amount: HALF},
-        {address: bar, amount: HALF},
-      ];
-      const expectedDistribution = createImmediateDistribution(
-        20,
-        expectedReceipts
-      );
-      expectDistributionsEqual(result, expectedDistribution);
-    });
-    it("handles an interval with un-even cred createGrainDistribution", () => {
-      const result = createGrainDistribution(
-        strategy,
-        credHistory,
-        new Map(),
-        12
-      );
-      // $ExpectFlowError
-      const ONE_TENTH = ONE / 10n;
-      const NINE_TENTHS = ONE - ONE_TENTH;
-      const expectedReceipts = [
-        {address: foo, amount: NINE_TENTHS},
-        {address: bar, amount: ONE_TENTH},
-      ];
-      const expectedDistribution = createImmediateDistribution(
-        12,
-        expectedReceipts
-      );
-      expectDistributionsEqual(result, expectedDistribution);
-    });
-    it("handles an interval at the end", () => {
-      const result = createGrainDistribution(
-        strategy,
-        credHistory,
-        new Map(),
-        1000
-      );
-      const expectedReceipts = [{address: bar, amount: ONE}];
-      const expectedDistribution = createImmediateDistribution(
-        1000,
-        expectedReceipts
-      );
-      expectDistributionsEqual(result, expectedDistribution);
-    });
-  });
-
-  describe("balancedDistribution", () => {
-    const strategy = deepFreeze({
-      type: "BALANCED",
-      budget: fromApproximateFloat(14),
-      version: 1,
-    });
-
-    const createBalancedDistribution = (
-      timestampMs: number,
-      receipts: $ReadOnlyArray<GrainReceipt>,
-      strategy: BalancedV1
-    ) => {
-      return {
-        type: "GRAIN_DISTRIBUTION",
-        version: GRAIN_DISTRIBUTION_VERSION_1,
-        receipts,
-        strategy,
-        timestampMs,
-      };
-    };
-
-    describe("it should return an empty distribution when", () => {
-      const emptyDistribution = deepFreeze({
-        type: "GRAIN_DISTRIBUTION",
-        version: GRAIN_DISTRIBUTION_VERSION_1,
-        receipts: [],
-        strategy,
-        timestampMs,
-      });
-
-      it("the budget is zero", () => {
-        const zeroBudgetStrategy = {...strategy, budget: ZERO};
-
-        const actual = createGrainDistribution(
-          zeroBudgetStrategy,
-          credHistory,
-          new Map(),
-          timestampMs
-        );
-
-        expectDistributionsEqual(actual, {
-          ...emptyDistribution,
-          strategy: zeroBudgetStrategy,
-        });
-      });
-
-      it("there are no cred scores at all", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          [],
-          new Map(),
-          timestampMs
-        );
-
-        expectDistributionsEqual(actual, emptyDistribution);
-      });
-
-      it("all cred scores are from the future", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          credHistory,
-          new Map(),
-          0
-        );
-        expectDistributionsEqual(actual, {
-          ...emptyDistribution,
-          timestampMs: 0,
-        });
-      });
-
-      it("all the cred sums to 0", () => {
-        const actual = createGrainDistribution(
-          strategy,
-          [
-            {
-              intervalEndMs: timestampMs - 500,
-              cred: new Map([
-                [foo, 0],
-                [bar, 0],
-              ]),
-            },
-          ],
-          new Map(),
-          timestampMs
-        );
-
-        expectDistributionsEqual(actual, emptyDistribution);
-      });
-    });
-
-    it("throws an error if given an unsupported strategy", () => {
-      const unsupportedStrategy = {
-        ...strategy,
-        version: 2,
-      };
-      expect(() =>
-        createGrainDistribution(
-          unsupportedStrategy,
-          credHistory,
-          new Map(),
-          timestampMs
-        )
-      ).toThrowError(`Unsupported BALANCED strategy: 2`);
-    });
-
-    it("should only pay Foo if Foo is sufficiently underpaid", () => {
-      const lifetimeEarnings = new Map([
-        [foo, ZERO],
-        [bar, fromApproximateFloat(99)],
-      ]);
-      const expectedReceipts = [
-        {address: foo, amount: fromApproximateFloat(14)},
-      ];
-      const expectedDistribution = createBalancedDistribution(
-        timestampMs,
-        expectedReceipts,
-        strategy
-      );
-      const actual = createGrainDistribution(
-        strategy,
-        credHistory,
-        lifetimeEarnings,
-        timestampMs
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-    it("should divide according to cred if everyone is already balanced paid", () => {
-      const lifetimeEarnings = new Map([
-        [foo, fromApproximateFloat(5)],
-        [bar, fromApproximateFloat(2)],
-      ]);
-
-      const expectedReceipts = [
-        {address: foo, amount: fromApproximateFloat(10)},
-        {address: bar, amount: fromApproximateFloat(4)},
-      ];
-
-      const expectedDistribution = createBalancedDistribution(
-        timestampMs,
-        expectedReceipts,
-        strategy
-      );
-
-      const actual = createGrainDistribution(
-        strategy,
-        credHistory,
-        lifetimeEarnings,
-        timestampMs
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-    it("'top off' users who were slightly underpaid'", () => {
-      // Foo is exactly 1 grain behind where they "should" be
-      const lifetimeEarnings = new Map([
-        [foo, fromApproximateFloat(4)],
-        [bar, fromApproximateFloat(2)],
-      ]);
-
-      const strategy15 = {...strategy, budget: fromApproximateFloat(15)};
-
-      const expectedReceipts = [
-        {address: foo, amount: fromApproximateFloat(11)},
-        {address: bar, amount: fromApproximateFloat(4)},
-      ];
-
-      const expectedDistribution = createBalancedDistribution(
-        timestampMs,
-        expectedReceipts,
-        strategy15
-      );
-
-      const actual = createGrainDistribution(
-        strategy15,
-        credHistory,
-        lifetimeEarnings,
-        timestampMs
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-    it("should ignore cred scores from the future", () => {
-      const middleTimes = 29;
-
-      const strategy12 = {...strategy, budget: fromApproximateFloat(12)};
-
-      // In the last time slice, foo gets 0 cred and bar gets 2
-      const expectedReceipts = [
-        {address: foo, amount: fromApproximateFloat(10)},
-        {address: bar, amount: fromApproximateFloat(2)},
-      ];
-
-      const expectedDistribution = createBalancedDistribution(
-        middleTimes,
-        expectedReceipts,
-        strategy12
-      );
-
-      const actual = createGrainDistribution(
-        strategy12,
-        credHistory,
-        new Map(),
-        middleTimes
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-    it("should handle the case where one user has no historical earnings", () => {
-      const lifetimeEarnings = new Map([[foo, fromApproximateFloat(5)]]);
-
-      const expectedReceipts = [
-        {address: bar, amount: fromApproximateFloat(2)},
-      ];
-      const strategy2 = {...strategy, budget: fromApproximateFloat(2)};
-
-      const expectedDistribution = createBalancedDistribution(
-        timestampMs,
-        expectedReceipts,
-        strategy2
-      );
-
-      const actual = createGrainDistribution(
-        strategy2,
-        credHistory,
-        lifetimeEarnings,
-        timestampMs
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-    it("should not break if a user has earnings but no cred", () => {
-      const lifetimeEarnings = new Map([
-        [NodeAddress.fromParts(["zoink"]), fromApproximateFloat(10)],
-      ]);
-
-      const expectedReceipts = [
-        {address: foo, amount: fromApproximateFloat(10)},
-        {address: bar, amount: fromApproximateFloat(4)},
-      ];
-
-      const expectedDistribution = createBalancedDistribution(
-        timestampMs,
-        expectedReceipts,
-        strategy
-      );
-
-      const actual = createGrainDistribution(
-        strategy,
-        credHistory,
-        lifetimeEarnings,
-        timestampMs
-      );
-      expectDistributionsEqual(expectedDistribution, actual);
-    });
-  });
-});
